middleware changes

+++ app/Http/Middleware/CleanupLivewireTempFiles.php
@@ -23,14 +23,19 @@ class CleanupLivewireTempFiles
     protected function cleanupOldLivewireTempFiles()
     {
         $tempDir = storage_path('app/livewire-tmp');
-        $files = File::files($tempDir);
 
-        foreach ($files as $file) {
-            $fileCreationTime = \Carbon\Carbon::createFromTimestamp($file->getCTime());
-            if (now()->diffInMinutes($fileCreationTime)>60) {
-                Log::info('Deleting old Livewire temp file:', ['file' => $file->getRealPath()]);
-                File::delete($file);
+        if (File::exists($tempDir)) {  // Check if the directory exists
+            $files = File::files($tempDir);
+
+            if (!empty($files)) {  // Check if there are any files in the directory
+                foreach ($files as $file) {
+                    $fileCreationTime = \Carbon\Carbon::createFromTimestamp($file->getCTime());
+                    if (now()->diffInMinutes($fileCreationTime) > 60) {
+                        File::delete($file);
+                    }
+                }
             }
         }
     }
+
 }
